requirements.txt
streamlit==1.23.0
requests==2.31.0
Pillow==9.5.0